Autosaves per Game Name - Quality of Life Enhancement (#16)
* Autosaves per Game Name - Quality of Life Enhancement
When I play multiplayer games, I often cycle between files frequently depending on who I'm playing with. I have a file here that I play with one friend, another there that I player with a couple others, so on and so forth. With all this switching around, it can be easy to forget to save a Replay and have all traces of that file overwritten as your last hope - the autosaves - are overwritten by the other games you're hosting.
This change solves that by including the name of the game session in the autosave's name. Each name will support the usual 5 slots as per normal. For example, if I host a game and give it the name "Together", then the autosaves created from this will be named "Autosave-Together-1", "Autosave-Together-2", etc... and will not overwrite other autosaves from sessions with other names.
Note: the autosave's use the name of the game session (the name picked in the host dialog), not the name of the loaded file/replay.

* Added toggle that controls whether or not the game name will be appended to the autosave names. The toggle is set by the host from the HostWindow and defaults to off to maintain the default user experience.
Added new entry to the language files for this new option. The same value is provided for each (in English) so these will need to be localized later.

* Moved the appendNameToAutosave option to the Mod Setting page.