Addes sync worker for PawnTab
This fixes the issue of being unable to toggle some options in animals and wildlife tabs, like slaughter/release animal or designating a wild animal for hunting/taming.

I've put it into the Pawns region, though as it seems to be mostly related to tabs, it could fit the Tabs regions as well.