Initial work on PersistentDialogs attached to letters
Currently this applies to all letters that open up any dialog inheriting from Dialog_NodeTree, which is not always what we want. There'll need to be added a list of types that it'll be applied to.

This also fixes dialogs not blocking the time when they're supposed to if there was no map.