Fix neuroquake ability
This also affects other abilities that require confirmation (due to def having confirmationDialogText filled, as opposed to confirmations from effect comps).

The change here is to call the compiler-generated method which starts casting the ability directly. Right now we're calling the method which checks if confirmation is needed first - so it worked fine in cases where there was no confirmation needed.

The dialog is only opened for the player that clicks the gizmo, and the action is started once the player accepts the dialog.

I've also moved the method registration from the SyncMethods class to SyncDelegates, as the synced methods are now using delegates.