Dont start a job to go to the styling station until confirming
Right now in MP, if you order a pawn to use a styling station they'll walk there, only to open the dialog for the player who ordered it, and walk away while that player is modifying their style.

I've modified the behaviour here to not take the job and instead open the dialog for the player who ordered it - the job to change style will be applied after accepting the dialog, like it was previously.