Prevent creation of TabStorageWrapper when storage is a thing/comp with null map (#293)
I'm unsure how this exactly happens, but it's happening consistently with Save our Ship 2 torpedoes.

I assumed it'd be better to patch it here instead of doing harmony prefixes to prevent it from being created, as it's (in my eyes) much cleaner solution to this issue. And in case it ever would appear in another mod, it'll help with those situations too.