Synced a significant portion of gizmos
This includes quite a bit of dev mode gizmos as well. I'm pretty sure I missed a bit of them.

I tested all the gizmos I could. However, a big portion of them are DLC content which, due to the DLC not being released yet, I was unable to test.

Two classes that I'm sure need their gizmos synced are HumanEmbryo and HumanOvum. I've left them for now as they're a bit more on the complex side.

There may be gizmos I've missed as well. But patching those will most likely have to wait for the actual release of the DLC.