Update to 1.4, pass 1 (#307)
* Initial work on 1.4

* Synced get relic to inventory command

* Fixed pawns not being able to rotate visually

* Updated About.xml

- increased the required version to 1.4.3509
- added Biotech in loadAfter (ID was included in About.xml for other DLCs)

* Synced extract relic to area

* Synced style selection dialog

* Synced rocketswarm launcher gizmo

* Synced a significant portion of gizmos

This includes quite a bit of dev mode gizmos as well. I'm pretty sure I missed a bit of them.

I tested all the gizmos I could. However, a big portion of them are DLC content which, due to the DLC not being released yet, I was unable to test.

Two classes that I'm sure need their gizmos synced are HumanEmbryo and HumanOvum. I've left them for now as they're a bit more on the complex side.

There may be gizmos I've missed as well. But patching those will most likely have to wait for the actual release of the DLC.

* Sync changing the color of the light

Also added a sync worker for ColorInt, as it was required for it

* Small fixes

Change CountToTransferChanged to Notify_TransferablesChanged, use CompAutoCut.DesignatePlantsToCut instead of CompAnimalPenMarker.DesignatePlantsToCut

* Fix Pawn.ProcessPostTickVisuals patch

* Move postTickVisuals.ProcessPostTickVisuals to ticking

* Fixed prisoner interaction tab

* Fixed trade window crash

* Fix the travel supplies text in caravan forming dialog

* Fixed operations

* Fixed public execution ritual

* Fixed accepting quests

The `quests` field was changed into `allQuests` field. The `QuestsListForReading` getter returns the correct field, no matter which version.

* Updated RimWorld reference to 1.4.3517

* Update Languages

* Fixed throne speech/leader speech abilities

* Fixed pawns being re-assigned as spectators in rituals when re-opening the dialog

This also includes pawns with forced roles being assigned as spectators, locking them there without the ability to change it

* Synced storage groups (storage shelves)

* Reimplemented the mod compatibility button in mod list

* Synced gene resource gizmo

* Synce dev mode need changes

* Slight SyncFields formatting change

* Fix fleck related RNG desync

* Fix more fleck RNG desyncs

* Add Things from ThingOwner to sharedCrossRefs, remove relic extraction sync

Since we add those to sharedCrossRefs, it's no longer needed for extracting relics. This also fixes the same issue with removing waste packs from mechanoid gestators.

* Synced mechanitor control groups

* Restore code that got lost in a merge

* Remove temporary code

I completely forgot that I set it up... Oops.

* Synced HumanEmbryo/HumanOvum gizmos

* Fix mechanitor recharge thresholds after merge

* Synced mech pawn column

* Synced sterilize animal option in pawn column

It was already getting synced, however by syncing the method itself it'll set the column as dirty and cause them to be sorted again (if they were set to be sorted).

* Sync cut all blighted gizmo

* Prevent no pawn with skill dialog from opening for players who didn't issue the bill

* Sync gene assembler dialog

* Sync xenogerm gizmos

* Update Rimworld dependency

* Synced mech color change

Co-authored-by: Sokyran <toCy&#$VIH$1cSOu8s&Fv@UVN%L3snN*H09$1^Xghq*hKo9Itnmq#bNf8Ld67#T$#a&cEaY1>