Added sync for IGeneResourceDrain
Unnecessary for MP with no other mods. Added for the purpose of mod compatibility.

All vanilla implementations of IGeneResourceDrain, as well as all modded ones I could find, are a subtype of the Gene class - so for implementation, I check if it's a subtype of it and throw an exception otherwise.