Remove 3 session interfaces in favor of abstract classes
Removed `ISession`, `IExposableSession`, and `ISemiPersistentSession` in favor of abstract classes `Session`, `ExposableSession`, and `SemiPersistentSession`.

There was likely no class besides `Session` that would implement `ISession`, so it was removed. The other 2 were removed in favor of turning them into abstract subclasses of `Session`, which will allow for expanding them further in the future, and prevent people from implementing both interfaces (which was not supported).

All the xml docs were moved from interfaces to their respective classes.