Sync Dialog_BillConfig target item style change
The feature is only available with Ideology DLC active while the ideo system is disabled.

All of the prefixed methods could be turned into sync methods/delegates, but it would require making a sync worker for Dialog_BillConfig or making an instance/field transformer for it, so it seemed using sync fields would end up simpler.