Added rule to load before Vanilla Expanded Framework
This is a minor change that shouldn't really affect anything. The main purpose of this is to try and enforce more proper load order, similarly to the reason we list HugsLib as a mod that we should load before.

HugsLib is no longer receiving support and contains some issues, and some mods are dropping it as dependency due to this. Meanwhile Vanilla Expanded Framework is seeing more and more use - be it from new Vanilla Expanded mods, or even unrelated mods that are using it for its features - is