IVerbOwner, ISelectable - support all implementations by default
I've placed `IVerbOwner` sync worker in the `Interfaces` region, which I've placed at the very bottom. This is due to issues with syncing `IVerbOwner` - for example, trying to sync a `Pawn` caused it to be synced as `IVerbOwner`, which caused it to be recursively synced as `IVerbOwner` until the game crashed.

Placing it lower fixed the issue.