Update Krafs.Rimworld.Ref NuGet package to 1.4.3901
Considering that in About.xml we mention that MP requires v1.4.3901, we may as well use NuGet package for that specific version.

This doesn't really change anything besides that, but should ensure we don't touch anything that may have been removed and that we'll be able to reference anything new in code.

The NuGet reference was a bit inconsistent across projects, since Multiplayer and the Loader were both using 1.4.3704, while Common was using 1.4.3551-beta.