Optimize operations on dictionaries, fix small issue
Occurrences of `ContainsKey` followed by indexer getter/setter were replaced by `TryGetValue`/`TryAdd`. In both cases they'll have the same effect.

This also fixes a minor issue with `TypeCache` where `typeByFullName` would check for existence of `Type.Name`, but insert `Type.FullName`.

For extra context/explanation behind dictionary operation changes:

`TryAdd`, `Add`, and indexer all use the same method for adding to dictionary (`TryInsert`, using a different `InsertionBehavior`). This also means that all of them do a check if a value exists in the dictionary before adding the value - and this is the only way they differ. When an entry with a given key exists `TryAdd` just returns early, `Add` will throw an exception, and indexer setter will replace it.

As for `TryGetValue` - it just consolidates the `ContainsKey` and indexer getter into a single call, as all 3 methods call `FindEntry`. This will drop the amount of `FindEntry` calls by 1 each time it's used.