Include if async time and multifaction are active in desync logs (#421)
As far as I'm aware, this information is currently not included in the desync logs. It was sometimes possible to determine if it was by analyzing log files or desync files, but it was not an ideal solution.

Having this info could occasionally prove useful, especially if person sharing the logs is currently not able to respond when asked if async/multifaction were active.