Added a check if CompPollutionPump current map patch failed
Added a simple check if patch for `CompPollutionPump.CompTick` to replace `Find.CurrentMap` call was replaced with a safe call. If the patch failed a warning will be logged.

This should give a heads-up if this gets fixed in a future vanilla release, meaning this patch is no longer needed.