Update RimWorld related dependencies
Dependencies update to the latest version (in solutions using those):
- `Lib.Harmony` from 2.2.2 to 2.3.3
- `Krafs.Rimworld.Ref`
  - Updated from 1.5.4055-beta to 1.5.4062 in `Multiplayer.csproj`
  - Updated from 1.4.3901 to 1.5.4062 in `Common.csproj` and `MultiplayerLoader.csproj`
- `Krafs.Publicizer` from 2.0.1 to 2.2.1
  - This includes changes done automatically when updating, namely adding `PrivateAssets` and `IncludeAssets`
- `Zetrith.Prepatcher` from 1.1.1 to 1.2.0

An additional (and required) change is that `PresistentDialogs` now uses an explicit call to `Extensions.GetDefaultValue`. This is done due to a name conflict with `AccessToolsExtensions:GetDefaultValue`. (Perhaps we should use Harmony's `GetDefaultValue`, as it also handles `null` and `typeof(void)` types?)