Remove a duplicate StorageGroup sync worker, merge both storage regions
It seems a second StorageGroup sync worker was added recently, but the old one was kept in. I've removed the old one since the new one seems to be better (as it doesn't rely on a `Thing`, just a map and ID).

On top of that, I've merged the two regions named Storage.