Sync ActivityGizmo/CompActivity, don't sync HarbingerTree.UpdateRoots
Syncing of `HarbingerTree.UpdateRoots` was removed, as the method would naturally be queued up during simulation (SpawnSetup) and called in as a long event, causing the call to be synced. If dev mode syncing was disabled it would end up never being called.

Synced following fields:
- `ActivityGizmo.targetValuePct`
- `CompActivity.suppressIfAbove`
- `CompActivity.suppressionEnabled`

This required creating a sync worker for `ActivityGizmo`.

The syncing here is pretty similar to syncing of `GeneGizmo_Resource`, `Gene_Resource`, and `Gene_Hemogen`. It would potentially be possible to make a more universal approach, but I could not think of one myself.

`SyncGeneGizmoResource` and `SyncActivityGizmoTarget` sync the same field (`Gizmo_Slider.targetValuePct`), but we can't register sync field targeting `Gizmo_Slider` as that type is not syncable.