Add EarlyPatch attribute to PawnTweener.TweenedPos patch
Zetrith provided the fix on Discord and pointed out that it seems to be inlining issue. It was the first time I've dealt with an issue like this, so I greatly appreciate it. I've spent several day investigating the issue and have just reached a dead end.

This should fix the issue with Vanilla Expanded Framework. Using this fix I was not able to reproduce the issue anymore.