Use a randomized seed for long events
Previously, we've used a constant seed for long events. This change will replace it with a randomized seed which is selected whenever a long event is queued.

This should allow for more randomness in long events, and is required by the labyrinth map generation to produce unique maps, instead of only having a single possible layout.