Psychic ritual syncing
I've decided to look into this over the past 2-3 days and have mostly finished it, however there's still a few things that are needed (so it's currently a draft). Things still needing work:
- Modify ritual prepatches to also work with psychic rituals, so assigning roles will be synced
- Add localization for `MpPsychicRitualSession`

Info on psychic ritual patches I've included:
- Heavily based on ritual patches (started by copy-pasting the code related to them)
- I've not used "PsychicRitualData" like how it's done with "RitualData" for normal rituals, as I've decided there wasn't as much data to handle
- I've synced the start call by overriding the "Start" method in ritual proxy
  - This can also be applied to normal rituals to remove a harmony patch
  - In case of normal rituals, we could also include a confirmation dialog to match vanilla behaviour
- Syncing of creating session was done by patching `PsychicRitualGizmo.InitializePsychicRitual`
  - As opposed to normal rituals that can be started in multiple places, this is the only location those can be created from
- I believe everything else should have comments, match behaviour of other session, or be self-explanatory
  - If explanation for anything else is needed, let me know, and I'll include it in code comments or on GitHub

A slight side note about dialogs with sessions (and styling station dialog) - perhaps we should make a single harmony patch to `Widgets.ButtonTextWorker` with a generalized way to use it? As it stands right now, we basically make two patches per session (prefix and postfix). Having a more generalized way to handle it may be useful in the future.