Patch WindManager.BaseWindSpeedAt hash code usage
Changes wind speed calculation to be equal on all clients.

This patch replaces `object.GetHashCode()` with a deterministic hash based on `map.Tile` when in multiplayer.

In version 1.5, `this.map.Tile` was used instead of `this.map.GetHashCode()`.

In version 1.6, `map.Tile` is no longer an `int`, but a `PlanetTile` object. However, it can still be implicitly converted to an `int`.

Using `GetHashCode()`in 1.6 introduces wind speed randomness across game loads.
This patch ensures consistency between all clients in multiplayer, but it removes the randomness.