Fix visual ThrowLavaSmoke desync by swapping CurrentMap
The `ThrowLavaSmoke ` method spawns a small visual effect. This method is guarded by `ShouldSpawnMotesAt`.

`ShouldSpawnMotesAt` uses `Find.CurrentMap` and compares it to the provided map. In multi-faction gameplay - and possibly in multi-map gameplay - `CurrentMap` is not updated while ticking.

Therefore, for the map that is not spawning the visual effect, the comparison returns `false`, while on the other map it returns `true`, leading to a desync.

In this commit, `CurrentMap` is temporarily swapped to the target map within the `ThrowLavaSmoke ` call to ensure proper behavior.