Make sure the player list is sent before any other related packets.
As soon as a connection's state becomes Playing, player list updates can
be sent to it. However, sending a player list update before sending the
player list itself is wrong, so make sure to send it before other
packets. The timing required to trigger this seems improbable, but I did
see issues that could be related to this.