Fix: Cancel Launch On World Map Not Closing For Non Issuer
Refactor to use the original method for StopTargeting.

I spent quite a while trying to find what in the source code causes the map to not close for the non-issuer. I found that, for the issuer, the wantedMode is set to None and for the issuer to planet, but I couldn’t find where this happens in the source code. So I just set it manually.