Simplifies Gravship Sessions Handling
This commit fixes loading session errors, and joining a client during takeoff, landing, or an active session is now possible.

When a user joins, the current session is canceled. This works well for the takeoff session. In the landing session, however, when the landing position is selected, the session is not recreated. As a result, the time can still be controlled. I think this is fine for now, but I’ll note it down for future change.

We now use separate takeoff and landing sessions.
The world session has been removed.
Exposable sessions have also been removed.